requirements.txt
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
docutils==0.21.2
filetype==1.2.0
firebase-admin==6.8.0
gcloud==0.18.3
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
jws==0.1.3
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-examples==2.3.1
Kivy-Garden==0.1.5
kivy_deps.gstreamer==0.3.4
kivy_deps.sdl2==0.8.0
kivymd==1.2.0
msgpack==1.1.0
oauth2client==4.1.3
pillow==10.4.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pypiwin32==223
Pyrebase4==4.8.0
python-jwt==2.0.1
pywin32==310
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9.1
six==1.17.0
uritemplate==4.1.1
urllib3==1.26.20
